requirements.txt
wxPython
requests
youtube-transcript-api
Markdown
python-dotenv
